refactor: move miscelaneous epoch APIs to epoch manager adapter (#7994)
based on top of #7992

cc https://github.com/near/nearcore/issues/6910
